Change len arg of getidle() from size_t to int since it is used as a length parameter to a "%.*s" printf format.  Quiets a compiler warning (both gcc and clang).

+++ libexec/rpc.rusersd/rusers_proc.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: rusers_proc.c,v 1.24 2013/11/13 15:24:21 deraadt Exp $	*/
+/*	$OpenBSD: rusers_proc.c,v 1.25 2016/08/16 04:48:04 tedu Exp $	*/
 
 /*-
  *  Copyright (c) 1993 John Brezak
@@ -71,7 +71,7 @@ extern int from_inetd;
 FILE *ufp;
 
 static long
-getidle(char *tty, size_t len)
+getidle(char *tty, int len)
 {
 	char devname[PATH_MAX];
 	struct stat st;
